+++ src/main/java/lotto/io/ConsoleManager.java
@@ -26,7 +26,7 @@ public class ConsoleManager {
         }
     }
 
-    public void printLottoCount(final long count) {
+    public void printLottoCount(final int count) {
         System.out.println("\n" + count + "개를 구매했습니다.");
     }
 }
